// types/supabase.ts
export interface definitions {
  bookmarks: {
    /**
     * Note:
     * This is a Primary Key.<pk/>
     */
    id: number;
    created_at?: string;
    uuid?: string;
    /**
     * Note:
     * This is a Foreign Key to `posts.uuid`.<fk table='posts' column='uuid'/>
     */
    post_uuid?: string;
    /**
     * Note:
     * This is a Foreign Key to `users.uuid`.<fk table='users' column='uuid'/>
     */
    user_uuid?: string;
  };
// ...
  comments: {
    /**
     * Note:
     * This is a Primary Key.<pk/>
     */
    id: number;
    created_at?: string;
    uuid?: string;
    text?: string;
    /**
     * Note:
     * This is a Foreign Key to `posts.uuid`.<fk table='posts' column='uuid'/>
     */
    post_uuid?: string;
    /**
     * Note:
     * This is a Foreign Key to `users.uuid`.<fk table='users' column='uuid'/>
     */
    user_uuid?: string;
    /**
     * Note:
     * This is a Foreign Key to `comments.uuid`.<fk table='comments' column='uuid'/>
     */
    parent_uuid?: string;
  };
// ...
  posts: {
    /**
     * Note:
     * This is a Primary Key.<pk/>
     */
    id: number;
    created_at?: string;
    title?: string;
    description?: string;
    type?: string;
    views?: number;
    options?: string;
    uuid?: string;
    /**
     * Note:
     * This is a Foreign Key to `users.uuid`.<fk table='users' column='uuid'/>
     */
    user: string;
    image?: string;
    url?: string;
    /**
     * Note:
     * This is a Foreign Key to `topics.uuid`.<fk table='topics' column='uuid'/>
     */
    topic?: string;
  };
// ...
}